Fix namespace handling for ConfigMapConfigProperties. Remove duplicate code between ConfigMapPropertySource and SecretsPropertySource.
