+++ spring-cloud-kubernetes-core/src/main/java/io/fabric8/spring/cloud/kubernetes/config/AbstractConfigProperties.java
@@ -17,15 +17,36 @@
 
 package io.fabric8.spring.cloud.kubernetes.config;
 
-import org.springframework.boot.context.properties.ConfigurationProperties;
 
-@ConfigurationProperties("spring.cloud.kubernetes.config")
-public class ConfigMapConfigProperties extends AbstractConfigProperties {
+public abstract class AbstractConfigProperties {
 
-    private static final String TARGET = "Config Map";
+    protected boolean enabled = true;
+    protected String name;
+    protected String namespace;
 
-    @Override
-    public String getConfigurationTarget() {
-        return TARGET;
+    public abstract String getConfigurationTarget();
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getNamespace() {
+        return namespace;
+    }
+
+    public void setNamespace(String namespace) {
+        this.namespace = namespace;
     }
 }
